update routings & move dependencies to core
