+++ clients/desktop/src/navigation/router.tsx
@@ -1,12 +1,15 @@
+import { AddressPage } from '@core/ui/chain/components/address/AddressPage'
 import { CorePath, corePaths } from '@core/ui/navigation'
 import { sharedRoutes } from '@core/ui/navigation/routes'
 import { IncompleteOnboardingOnly } from '@core/ui/onboarding/IncompleteOnboardingOnly'
 import { ActiveVaultGuard } from '@core/ui/vault/ActiveVaultGuard'
+import { VaultChainCoinPage } from '@core/ui/vault/chain/coin/VaultChainCoinPage'
+import { ManageVaultChainCoinsPage } from '@core/ui/vault/chain/manage/coin/ManageVaultChainCoinsPage'
+import { VaultChainPage } from '@core/ui/vault/chain/VaultChainPage'
 import { toEntries } from '@lib/utils/record/toEntries'
 import { ReactNode } from 'react'
 import { createBrowserRouter, Outlet } from 'react-router-dom'
 
-import { AddressPage } from '../chain/components/address/AddressPage'
 import { DeeplinkPage } from '../deeplink/components/DeeplinkPage'
 import { ErrorBoundary } from '../errors/components/ErrorBoundary'
 import { FullSizeErrorFallback } from '../errors/components/FullSizeErrorFallback'
@@ -18,9 +21,6 @@ import SettingsVaultPage from '../pages/vaultSettings/SettingsVaultPage'
 import AddressBookSettingsPage from '../pages/vaultSettings/vaultAddressBook/AddressBookSettingsPage'
 import VaultCheckUpdatePage from '../pages/vaultSettings/vaultCheckUpdatePage/VaultCheckUpdatePage'
 import FaqVaultPage from '../pages/vaultSettings/vaultFaq/FaqVaultPage'
-import { VaultChainCoinPage } from '../vault/chain/coin/VaultChainCoinPage'
-import { ManageVaultChainCoinsPage } from '../vault/chain/manage/coin/ManageVaultChainCoinsPage'
-import { VaultChainPage } from '../vault/chain/VaultChainPage'
 import { VaultPage } from '../vault/components/VaultPage'
 import { DepositPage } from '../vault/deposit/DepositPage'
 import EditVaultPage from '../vault/edit/EditVaultPage'
@@ -91,6 +91,27 @@ const coreRoutes: Record<CorePath, ReactNode> = {
   ),
   uploadQr: <UploadQrPage />,
   vaults: <VaultsPage />,
+  vaultChainDetail: (
+    <ActiveVaultGuard>
+      <VaultChainPage />
+    </ActiveVaultGuard>
+  ),
+  vaultChainCoinDetail: (
+    <ActiveVaultGuard>
+      <VaultChainCoinPage />
+    </ActiveVaultGuard>
+  ),
+  manageVaultChainCoins: (
+    <ActiveVaultGuard>
+      <ManageVaultChainCoinsPage />
+    </ActiveVaultGuard>
+  ),
+  address: <AddressPage />,
+  deposit: (
+    <ActiveVaultGuard>
+      <DepositPage />
+    </ActiveVaultGuard>
+  ),
 }
 
 const appRoutes: Record<AppPath, ReactNode> = {
@@ -105,21 +126,6 @@ const appRoutes: Record<AppPath, ReactNode> = {
   shareVault: (
     <ActiveVaultGuard>
       <ShareVaultPage />
-    </ActiveVaultGuard>
-  ),
-  manageVaultChainCoins: (
-    <ActiveVaultGuard>
-      <ManageVaultChainCoinsPage />
-    </ActiveVaultGuard>
-  ),
-  vaultChainDetail: (
-    <ActiveVaultGuard>
-      <VaultChainPage />
-    </ActiveVaultGuard>
-  ),
-  vaultChainCoinDetail: (
-    <ActiveVaultGuard>
-      <VaultChainCoinPage />
     </ActiveVaultGuard>
   ),
   editVault: (
@@ -137,7 +143,6 @@ const appRoutes: Record<AppPath, ReactNode> = {
       <DeleteVaultPage />
     </ActiveVaultGuard>
   ),
-  address: <AddressPage />,
   vaultFAQ: <FaqVaultPage />,
   addressBook: <AddressBookSettingsPage />,
   migrateVault: (
@@ -168,11 +173,6 @@ const appRoutes: Record<AppPath, ReactNode> = {
       <CurrentVaultFolderPageProvider>
         <ManageVaultFolderPage />
       </CurrentVaultFolderPageProvider>
-    </ActiveVaultGuard>
-  ),
-  deposit: (
-    <ActiveVaultGuard>
-      <DepositPage />
     </ActiveVaultGuard>
   ),
   deeplink: (
