+++ clients/desktop/src/pages/vaultSettings/SettingsVaultPage.tsx
@@ -18,6 +18,7 @@ import { ShareIcon } from '@lib/ui/icons/ShareIcon'
 import ShieldCheckIcon from '@lib/ui/icons/ShieldCheckIcon'
 import TwitterIcon from '@lib/ui/icons/TwitterIcon'
 import { VultisigLogoIcon } from '@lib/ui/icons/VultisigLogoIcon'
+import { ScrollableFlexboxFiller } from '@lib/ui/layout/ScrollableFlexboxFiller'
 import { HStack, VStack } from '@lib/ui/layout/Stack'
 import { PageHeader } from '@lib/ui/page/PageHeader'
 import { PageHeaderBackButton } from '@lib/ui/page/PageHeaderBackButton'
@@ -27,7 +28,6 @@ import { Text } from '@lib/ui/text'
 import { ReactNode } from 'react'
 import { useTranslation } from 'react-i18next'
 
-import { ScrollableFlexboxFiller } from '../../lib/ui/layout/ScrollableFlexboxFiller'
 import { NavigateToDklsPage } from '../../mpc/dkls/NavigateToDklsPage'
 import { useAppNavigate } from '../../navigation/hooks/useAppNavigate'
 import {
