+++ clients/desktop/src/vault/balance/VaultTotalBalance.tsx
@@ -1,5 +1,7 @@
 import { fiatCurrencySymbolRecord } from '@core/config/FiatCurrency'
 import { useFiatCurrency } from '@core/ui/storage/fiatCurrency'
+import { BalanceVisibilityAware } from '@core/ui/vault/balance/visibility/BalanceVisibilityAware'
+import { useVaultTotalBalanceQuery } from '@core/ui/vault/queries/useVaultTotalBalanceQuery'
 import { HStack } from '@lib/ui/layout/Stack'
 import { Spinner } from '@lib/ui/loaders/Spinner'
 import { MatchQuery } from '@lib/ui/query/components/MatchQuery'
@@ -7,8 +9,6 @@ import { Text } from '@lib/ui/text'
 import { formatAmount } from '@lib/utils/formatAmount'
 import { useTranslation } from 'react-i18next'
 
-import { useVaultTotalBalanceQuery } from '../queries/useVaultTotalBalanceQuery'
-import { BalanceVisibilityAware } from './visibility/BalanceVisibilityAware'
 import { ManageVaultBalanceVisibility } from './visibility/ManageVaultBalanceVisibility'
 
 export const VaultTotalBalance = () => {
