--- clients/desktop/src/vault/chain/useCurrentVaultChain.tsx
@@ -1,7 +1,0 @@
-import { useAppPathParams } from '../../navigation/hooks/useAppPathParams'
-
-export const useCurrentVaultChain = () => {
-  const [{ chain }] = useAppPathParams<'vaultChainDetail'>()
-
-  return chain
-}
