+++ clients/desktop/src/vault/components/VaultPage.tsx
@@ -6,13 +6,13 @@ import { VStack } from '@lib/ui/layout/Stack'
 import { pageConfig } from '@lib/ui/page/config'
 import { PageHeader } from '@lib/ui/page/PageHeader'
 import { PageHeaderIconButton } from '@lib/ui/page/PageHeaderIconButton'
+import { PageHeaderIconButtons } from '@lib/ui/page/PageHeaderIconButtons'
 import { Link } from 'react-router-dom'
 import styled from 'styled-components'
 
 import UpdateAvailablePopup from '../../components/updateAvailablePopup/UpdateAvailablePopup'
 import { makeAppPath } from '../../navigation'
 import { PageHeaderVaultSettingsPrompt } from '../../pages/vaultSettings/PageHeaderVaultSettingsPrompt'
-import { PageHeaderIconButtons } from '../../ui/page/PageHeaderIconButtons'
 import { PageHeaderToggleTitle } from '../../ui/page/PageHeaderToggleTitle'
 import { RefreshVaultBalance } from '../../vault/balance/RefreshVaultBalance'
 import { VaultOverview } from '../../vault/components/VaultOverview'
