+++ clients/desktop/src/vault/deposit/ChainAction.ts
@@ -1,6 +1,5 @@
 import { Chain } from '@core/chain/Chain'
-
-import { DepositEnabledChain } from './DepositEnabledChain'
+import { DepositEnabledChain } from '@core/ui/vault/deposit/DepositEnabledChain'
 
 export type ChainAction =
   | 'bond'
