+++ clients/desktop/src/vault/deposit/DepositPageController.tsx
@@ -1,20 +1,20 @@
 import { coinKeyFromString } from '@core/chain/coin/Coin'
+import { useCorePathParams } from '@core/ui/navigation/hooks/useCorePathParams'
+import { DepositEnabledChain } from '@core/ui/vault/deposit/DepositEnabledChain'
 import { Match } from '@lib/ui/base/Match'
 import { useStepNavigation } from '@lib/ui/hooks/useStepNavigation'
 import { useNavigateBack } from '@lib/ui/navigation/hooks/useNavigateBack'
 import { useState } from 'react'
 import { FieldValues } from 'react-hook-form'
 
-import { useAppPathParams } from '../../navigation/hooks/useAppPathParams'
 import { ChainAction, chainActionsRecord } from './ChainAction'
-import { DepositEnabledChain } from './DepositEnabledChain'
 import { DepositForm } from './DepositForm'
 import { DepositVerify } from './DepositVerify'
 
 const depositSteps = ['form', 'verify'] as const
 
 export const DepositPageController = () => {
-  const [{ coin: coinName }] = useAppPathParams<'deposit'>()
+  const [{ coin: coinName }] = useCorePathParams<'deposit'>()
   const { chain: chain } = coinKeyFromString(coinName)
   const chainActionOptions = chainActionsRecord[chain as DepositEnabledChain]
