+++ clients/extension/src/navigation/router.tsx
@@ -1,5 +1,3 @@
-import { VaultChainCoinPage } from '@clients/desktop/src/vault/chain/coin/VaultChainCoinPage'
-import { VaultChainPage } from '@clients/desktop/src/vault/chain/VaultChainPage'
 import { OnboardingPage } from '@clients/extension/src/components/onboarding/components/OnboardingPage'
 import { ReshareFastVault } from '@clients/extension/src/components/settings/reshare/ReshareFastVault'
 import { ReshareSecureVault } from '@clients/extension/src/components/settings/reshare/ReshareSecureVault'
@@ -19,14 +17,17 @@ import { TransactionPage } from '@clients/extension/src/pages/transaction'
 import { VaultPage } from '@clients/extension/src/pages/vault'
 import { VaultSettingsPage } from '@clients/extension/src/pages/vault-settings'
 import { VaultsPage } from '@clients/extension/src/pages/vaults'
+import { AddressPage } from '@core/ui/chain/components/address/AddressPage'
 import { CorePath, corePaths } from '@core/ui/navigation'
 import { sharedRoutes } from '@core/ui/navigation/routes'
 import { IncompleteOnboardingOnly } from '@core/ui/onboarding/IncompleteOnboardingOnly'
 import { ActiveVaultGuard } from '@core/ui/vault/ActiveVaultGuard'
+import { VaultChainCoinPage } from '@core/ui/vault/chain/coin/VaultChainCoinPage'
+import { ManageVaultChainCoinsPage } from '@core/ui/vault/chain/manage/coin/ManageVaultChainCoinsPage'
+import { VaultChainPage } from '@core/ui/vault/chain/VaultChainPage'
 import { toEntries } from '@lib/utils/record/toEntries'
 import { ReactNode } from 'react'
 import { createHashRouter } from 'react-router-dom'
-
 const coreRoutes: Record<CorePath, ReactNode> = {
   ...sharedRoutes,
   vault: (
@@ -60,6 +61,27 @@ const coreRoutes: Record<CorePath, ReactNode> = {
       <VaultsPage />
     </ActiveVaultGuard>
   ),
+  vaultChainDetail: (
+    <ActiveVaultGuard>
+      <VaultChainPage />
+    </ActiveVaultGuard>
+  ),
+  vaultChainCoinDetail: (
+    <ActiveVaultGuard>
+      <VaultChainCoinPage />
+    </ActiveVaultGuard>
+  ),
+  manageVaultChainCoins: (
+    <ActiveVaultGuard>
+      <ManageVaultChainCoinsPage />
+    </ActiveVaultGuard>
+  ),
+  address: <AddressPage />,
+  deposit: (
+    <ActiveVaultGuard>
+      <>{/* <>TODO: Implement Deposit page</> */}</>
+    </ActiveVaultGuard>
+  ),
 }
 
 const appRoutes = {
@@ -88,16 +110,6 @@ const appRoutes = {
   importTab: <ImportFilePage />,
   vaultsTab: <GetVaultsPage />,
   transactionTab: <TransactionPage />,
-  vaultChainDetail: (
-    <ActiveVaultGuard>
-      <VaultChainPage />
-    </ActiveVaultGuard>
-  ),
-  vaultChainCoinDetail: (
-    <ActiveVaultGuard>
-      <VaultChainCoinPage />
-    </ActiveVaultGuard>
-  ),
 }
 
 export const router = createHashRouter(
