+++ core/ui/vault/balance/visibility/BalanceVisibilityAware.tsx
