+++ core/ui/vault/chain/coin/VaultChainCoinPage.tsx
@@ -2,24 +2,24 @@ import { useCoinPriceQuery } from '@core/ui/chain/coin/price/queries/useCoinPric
 import { useBalanceQuery } from '@core/ui/chain/coin/queries/useBalanceQuery'
 import { getBalanceQueryKey } from '@core/ui/chain/coin/queries/useBalancesQuery'
 import { adjustVaultChainCoinsLogos } from '@core/ui/vault/chain/manage/coin/adjustVaultChainCoinsLogos'
+import { VaultChainCoinItem } from '@core/ui/vault/chain/VaultChainCoinItem'
+import { VaultPrimaryActions } from '@core/ui/vault/components/VaultPrimaryActions'
 import { useCurrentVaultCoin } from '@core/ui/vault/state/currentVaultCoins'
 import { RefreshIcon } from '@lib/ui/icons/RefreshIcon'
+import { Center } from '@lib/ui/layout/Center'
 import { VStack } from '@lib/ui/layout/Stack'
 import { Spinner } from '@lib/ui/loaders/Spinner'
 import { PageContent } from '@lib/ui/page/PageContent'
 import { PageHeader } from '@lib/ui/page/PageHeader'
 import { PageHeaderBackButton } from '@lib/ui/page/PageHeaderBackButton'
 import { PageHeaderIconButton } from '@lib/ui/page/PageHeaderIconButton'
+import { PageHeaderIconButtons } from '@lib/ui/page/PageHeaderIconButtons'
 import { PageHeaderTitle } from '@lib/ui/page/PageHeaderTitle'
 import { Panel } from '@lib/ui/panel/Panel'
 import { MatchQuery } from '@lib/ui/query/components/MatchQuery'
 import { useInvalidateQueriesMutation } from '@lib/ui/query/hooks/useInvalidateQueriesMutation'
 import { useTranslation } from 'react-i18next'
 
-import { Center } from '../../../lib/ui/layout/Center'
-import { PageHeaderIconButtons } from '../../../ui/page/PageHeaderIconButtons'
-import { VaultPrimaryActions } from '../../components/VaultPrimaryActions'
-import { VaultChainCoinItem } from '../VaultChainCoinItem'
 import { useCurrentVaultCoinKey } from './useCurrentVaultCoinKey'
 
 export const VaultChainCoinPage = () => {
