+++ core/ui/vault/chain/manage/ManageVaultChainsPrompt.tsx
@@ -1,8 +1,7 @@
 import { makeCorePath } from '@core/ui/navigation'
+import { ListAddButton } from '@lib/ui/list/ListAddButton'
 import { useTranslation } from 'react-i18next'
 import { Link } from 'react-router-dom'
-
-import { ListAddButton } from '../../../lib/ui/list/ListAddButton'
 
 export const ManageVaultChainsPrompt = () => {
   const { t } = useTranslation()
