+++ core/ui/vault/chain/manage/coin/ManageVaultChainCoin.tsx
