+++ core/ui/vault/chain/manage/coin/ManageVaultChainCoinsPage.tsx
@@ -1,5 +1,6 @@
 import { CoinSearch } from '@core/ui/vault/chain/manage/coin/search/CoinSearch'
 import { useScrollTo } from '@lib/ui/hooks/useScrollTo'
+import { ScrollableFlexboxFiller } from '@lib/ui/layout/ScrollableFlexboxFiller'
 import { VStack } from '@lib/ui/layout/Stack'
 import { PageContent } from '@lib/ui/page/PageContent'
 import { PageHeader } from '@lib/ui/page/PageHeader'
@@ -9,7 +10,6 @@ import { PageSlice } from '@lib/ui/page/PageSlice'
 import { CurrentSearchProvider } from '@lib/ui/search/CurrentSearchProvider'
 import { useTranslation } from 'react-i18next'
 
-import { ScrollableFlexboxFiller } from '../../../../lib/ui/layout/ScrollableFlexboxFiller'
 import { VaultChainCoinOptions } from './search/VaultChainCoinOptions'
 
 export const ManageVaultChainCoinsPage = () => {
