+++ core/ui/vault/chain/useCurrentVaultChain.tsx
@@ -1,0 +1,7 @@
+import { useCorePathParams } from '../../navigation/hooks/useCorePathParams'
+
+export const useCurrentVaultChain = () => {
+  const [{ chain }] = useCorePathParams<'vaultChainDetail'>()
+
+  return chain
+}
