+++ core/ui/vault/components/VaultPrimaryActions/index.tsx
@@ -1,5 +1,6 @@
 import { CoinKey } from '@core/chain/coin/Coin'
 import { swapEnabledChains } from '@core/chain/swap/swapEnabledChains'
+import { DepositPrompt } from '@core/ui/vault/components/DepositPrompts'
 import { SendPrompt } from '@core/ui/vault/send/SendPrompt'
 import { useCurrentVaultNativeCoins } from '@core/ui/vault/state/currentVaultCoins'
 import { SwapPrompt } from '@core/ui/vault/swap/components/SwapPrompt'
@@ -7,8 +8,6 @@ import { UniformColumnGrid } from '@lib/ui/css/uniformColumnGrid'
 import { ValueProp } from '@lib/ui/props'
 import { isEmpty } from '@lib/utils/array/isEmpty'
 import { isOneOf } from '@lib/utils/array/isOneOf'
-
-import { DepositPrompt } from '../DepositPrompts'
 
 export const VaultPrimaryActions = ({ value }: Partial<ValueProp<CoinKey>>) => {
   const nativeCoins = useCurrentVaultNativeCoins()
