+++ core/ui/vault/queries/useVaultTotalBalanceQuery.ts
